refac: 메인/커뮤니티 페이지의 공통 컴포넌트 CommonPage에 합쳐 정리
