--- src/frontend/src/components/templates/MainPage.stories.tsx
@@ -1,8 +1,0 @@
-import MainPage from "./MainPage";
-
-export default {
-  title: "templates",
-  component: MainPage,
-};
-
-export const Main = () => <MainPage />;
