+++ src/frontend/src/pages/Common.tsx
@@ -1,0 +1,43 @@
+import PageContainer from "@components/atoms/Div/PageContainer";
+import HeaderHelmet from "@components/atoms/Helmet";
+import CommunitySettingModal from "@components/organisms/Modal/CommunitySettingModal";
+import CreateCommunityModal from "@components/organisms/Modal/CreateCommunityModal";
+import InviteFriendModal from "@components/organisms/Modal/InviteFriendModal";
+import UserSettingModal from "@components/organisms/Modal/UserSettingModal";
+import CommonPage from "@components/templates/CommonPage";
+import useModalStore from "@store/useModalStore";
+import { useEffect } from "react";
+import { useParams } from "react-router-dom";
+
+const modalTable = {
+  inviteFriend: <InviteFriendModal />,
+  userSetting: <UserSettingModal />,
+  communitySetting: <CommunitySettingModal />,
+  createCommunity: <CreateCommunityModal />,
+};
+
+const Common = () => {
+  const { communityId, channelId } = useParams();
+  const { setShowModal, showModal, modalType } = useModalStore();
+
+  const isMainPage = !communityId;
+  const channelName = "자바스크립트 스터디";
+
+  useEffect(() => {
+    setShowModal(false);
+  }, []);
+
+  const component = modalType ? modalTable[modalType] : <></>;
+
+  return (
+    <>
+      <HeaderHelmet title={`${isMainPage ? "친구" : channelName} | Discord`} />
+      <PageContainer>
+        <CommonPage isMainPage={isMainPage} />
+        {showModal && component}
+      </PageContainer>
+    </>
+  );
+};
+
+export default Common;
